ocr-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
opencv-python-headless==4.8.1.78
pytesseract==0.3.10
numpy==1.24.3
pydantic==2.5.0
python-dateutil==2.8.2
fuzzywuzzy==0.18.0
python-Levenshtein==0.21.1
